// Backend/src/domain/common/errors/enums/EErrorCode.ts
export enum EErrorCode {
  INTERNAL_ERROR = 'INTERNAL_ERROR',
  VALIDATION_ERROR = 'VALIDATION_ERROR',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',
  INVALID_EMAIL = 'INVALID_EMAIL',
  INVALID_PASSWORD = 'INVALID_PASSWORD',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  REGISTRATION_CLOSED = 'REGISTRATION_CLOSED',
  PASSWORD_MISMATCH = 'PASSWORD_MISMATCH',
  PASSWORD_TOO_SHORT = 'PASSWORD_TOO_SHORT',
  NAME_REQUIRED = 'NAME_REQUIRED',
  EMAIL_REQUIRED = 'EMAIL_REQUIRED',
  PASSWORD_REQUIRED = 'PASSWORD_REQUIRED',
  CONF_PASSWORD_REQUIRED = 'CONF_PASSWORD_REQUIRED',
  NO_FIELDS_TO_UPDATE = 'NO_FIELDS_TO_UPDATE',
  USER_NOT_IDENTIFIED = 'USER_NOT_IDENTIFIED',
  INVALID_TOKEN = 'INVALID_TOKEN',
  TOKEN_NOT_PROVIDED = 'TOKEN_NOT_PROVIDED',
  ACCESS_DENIED = 'ACCESS_DENIED',
  INVALID_OBJECT_ID = 'INVALID_OBJECT_ID',
  CAIXA_NOT_FOUND = 'CAIXA_NOT_FOUND',
  CAIXA_ALREADY_EXISTS = 'CAIXA_ALREADY_EXISTS',
  CAIXA_NAME_REQUIRED = 'CAIXA_NAME_REQUIRED',
  CAIXA_VALUE_INVALID = 'CAIXA_VALUE_INVALID',
  LANCAMENTO_NOT_FOUND = 'LANCAMENTO_NOT_FOUND',
  LANCAMENTO_VALUE_INVALID = 'LANCAMENTO_VALUE_INVALID',
  LANCAMENTO_DISTRIBUICAO_REQUIRED = 'LANCAMENTO_DISTRIBUICAO_REQUIRED',
  LANCAMENTO_DISTRIBUICAO_INVALID = 'LANCAMENTO_DISTRIBUICAO_INVALID',
  LANCAMENTO_CAIXA_ORIGEM_REQUIRED = 'LANCAMENTO_CAIXA_ORIGEM_REQUIRED',
  LANCAMENTO_COMPENSACAO_REQUIRED = 'LANCAMENTO_COMPENSACAO_REQUIRED',
  ORCAMENTO_VALUE_INVALID = 'ORCAMENTO_VALUE_INVALID',
  COMPETENCIA_INVALID = 'COMPETENCIA_INVALID',
  CAIXA_TIPO_INVALID = 'CAIXA_TIPO_INVALID',
  CAIXA_META_INVALID = 'CAIXA_META_INVALID',
  CAIXA_APORTE_INVALID = 'CAIXA_APORTE_INVALID',
  CAIXA_ORCAMENTO_INVALID = 'CAIXA_ORCAMENTO_INVALID',
}